+++ src/components/Camera/CameraComponent2.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import useCamera from './useCamera';
+import useCamera from './useCamera2';
 import {
 	CameraContainer,
 	VideoContainer,
