+++ src/components/Camera/CameraComponent3.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import useCamera from './useCamera';
+import useCamera from './useCamera3';
 import {
 	CameraContainer,
 	VideoContainer,
