+++ src/components/Camera/CameraComponent4.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import useCamera from './useCamera';
+import useCamera from './useCamera4';
 import {
 	CameraContainer,
 	VideoContainer,
